fix: Repaired both CI failures while preserving the narrow MetaMask Mobile telemetry filter behavior
Sentry-Issue: 6529-FRONTEND-AR
